Fix remaining clippy warnings across all crates
- Fix type mismatch in cam-binder metadata.rs (value_offset is u32)
- Add #[allow(unused_imports)] for conditional imports in cam-onnx
- Add Safety docs to unsafe functions in cam-hal-android
- Remove unnecessary casts (u16, isize) in cam-hal-android
- Use or_default() instead of or_insert_with(Vec::new)
- Simplify identical if-else blocks
- Add v4l2 feature flag to cam-hal-android Cargo.toml
- Remove unnecessary map_err identity functions in cam-binder/aidl.rs

All tests pass, zero clippy warnings across workspace.

+++ cam-rust/cam-hal-android/src/util.rs
@@ -163,7 +163,7 @@ pub fn unpack_raw10_to_16le(
             for (j, pix) in [pix0, pix1, pix2, pix3].iter().enumerate() {
                 let idx = out_row + (x + j) * 2;
                 if idx + 2 <= out.len() {
-                    let v = (*pix as u16) << 6; // left-justify to 16 bits
+                    let v = *pix << 6; // left-justify to 16 bits
                     out[idx..idx + 2].copy_from_slice(&v.to_le_bytes());
                 }
             }
